chore(utils): move BigIntMath in utils and update all refs
