+++ packages/cardano-services/src/tsconfig.json
@@ -7,6 +7,7 @@
   "references": [
     { "path": "../../core/src" },
     { "path": "../../ogmios/src" },
-    { "path": "../../rabbitmq/src" }
+    { "path": "../../rabbitmq/src" },
+    { "path": "../../util/src" }
   ]
 }
