+++ packages/cip2/src/tsconfig.json
@@ -4,7 +4,5 @@
     "outDir": "../dist/",
     "rootDir": "."
   },
-  "references": [
-    { "path": "../../core/src" }
-  ]
+  "references": [{ "path": "../../core/src" }, { "path": "../../util/src" }]
 }
