+++ packages/core/package.json
@@ -26,6 +26,7 @@
   "dependencies": {
     "@cardano-ogmios/client": "5.1.0",
     "@cardano-ogmios/schema": "5.1.0",
+    "@cardano-sdk/util": "0.2.0",
     "@emurgo/cardano-serialization-lib-browser": "9.1.0",
     "@emurgo/cardano-serialization-lib-nodejs": "9.1.0",
     "serialize-error": "^8",
