+++ packages/core/src/tsconfig.json
@@ -3,5 +3,6 @@
   "compilerOptions": {
     "outDir": "../dist/",
     "rootDir": "."
-  }
+  },
+  "references": [{ "path": "../../util/src" }]
 }
