+++ packages/core/src/util/index.ts
@@ -1,4 +1,3 @@
-export * from './BigIntMath';
 export * as util from './misc';
 export * from './slotCalc';
 export * from './txInspector';
