+++ packages/golden-test-generator/package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@cardano-ogmios/client": "4.1.0",
+    "@cardano-sdk/util": "0.2.0",
     "chalk": "^4.1.1",
     "clear": "^0.1.0",
     "cli-progress": "^3.9.0",
