+++ packages/util/src/BigIntMath.ts
