+++ packages/wallet/src/services/BalanceTracker.ts
@@ -1,5 +1,7 @@
 import { Balance, DelegationTracker, StakeKeyStatus, TransactionalObservables, TransactionalTracker } from './types';
-import { BigIntMath, Cardano, ProtocolParametersRequiredByWallet } from '@cardano-sdk/core';
+import { BigIntMath } from '@cardano-sdk/util';
+import { Cardano, ProtocolParametersRequiredByWallet } from '@cardano-sdk/core';
+
 import { Observable, combineLatest, distinctUntilChanged, map } from 'rxjs';
 import { TrackerSubject } from '@cardano-sdk/util-rxjs';
 import { deepEquals } from './util';
